backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
psycopg[binary]==3.1.18
pydantic[email]==2.6.1
pydantic-settings==2.2.1
reportlab==4.1.0
qrbill==1.2.0
svglib==1.5.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
sqlalchemy==2.0.27
